Fix Magick.NET resize argument types

+++ ResizeNfpsGifDialog.cs
@@ -137,8 +137,8 @@ namespace GifProcessorApp
             using var collection = new MagickImageCollection(input);
             foreach (var frame in collection)
             {
-                frame.Resize(width, height);
-                frame.AnimationDelay = fps > 0 ? (uint)Math.Round(100.0 / fps) : frame.AnimationDelay;
+                frame.Resize((uint)width, (uint)height);
+                frame.AnimationDelay = fps > 0 ? (int)Math.Round(100.0 / fps) : frame.AnimationDelay;
             }
             collection.Write(output);
         }
